--- a/error_private.h
+++ b/error_private.h
@@ -14,11 +14,6 @@
 #ifndef ERROR_H_MODULE
 #define ERROR_H_MODULE
 
-#if defined (__cplusplus)
-extern "C" {
-#endif
-
-
 /* ****************************************
 *  Dependencies
 ******************************************/
@@ -26,7 +21,6 @@ extern "C" {
 #include "compiler.h"
 #include "debug.h"
 #include "zstd_deps.h"       /* size_t */
-
 
 /* ****************************************
 *  Compiler-specific
@@ -161,10 +155,6 @@ void _force_has_format_string(const char *format, ...) {
             return err_code;                                                       \
         }                                                                          \
     } while(0)
-
-#if defined (__cplusplus)
-}
-#endif
 
 #endif /* ERROR_H_MODULE */