--- a/zstd_v02.c
+++ b/zstd_v02.c
@@ -2723,7 +2723,7 @@ typedef struct {
     BYTE* matchLength;
     BYTE* dumpsStart;
     BYTE* dumps;
-} seqStore_t;
+} SeqStore_t;
 
 
 /* *************************************