Same workflow as https://github.com/DataDog/zstd/pull/143

- [x] Run below:
```
git clone https://github.com/facebook/zstd.git c
cd c
git checkout v1.5.7
cd ..
git clone git@github.com:DataDog/zstd.git go
find c/lib \( -name '*.h' -o -name '*.c' -o -name '*.S' \) -exec cp -v {} go/ \;
```

- [x] Run `python tools/flatten_imports.py`
- [x] Run `python tools/insert_libzstd_ifdefs.py`


This is the code as of https://github.com/facebook/zstd/releases/tag/v1.5.7 (https://github.com/facebook/zstd/commit/f8745da6ff1ad1e7bab384bd1f9d742439278e99)

Benchmarks will come in the next comment