@Viq111 2025-03-28
Benchmarks on a Apple M1 Max:

```
goos: darwin
goarch: arm64
pkg: github.com/DataDog/zstd
cpu: Apple M1 Max
                                         │     old     │                new                 │
                                         │   sec/op    │   sec/op     vs base               │
BulkCompress-10                            891.4n ± 4%   899.2n ± 2%       ~ (p=0.481 n=10)
BulkDecompress-10                          415.3n ± 0%   411.1n ± 0%  -1.02% (p=0.000 n=10)
CtxCompression-10                          65.94m ± 0%   67.12m ± 0%  +1.79% (p=0.000 n=10)
CtxDecompression-10                        8.552m ± 0%   8.585m ± 0%  +0.38% (p=0.000 n=10)
SmallWriteStreamCompression/all-zeros-10   65.50n ± 0%   66.05n ± 0%  +0.84% (p=0.000 n=10)
SmallWriteStreamCompression/count-10       65.48n ± 0%   66.08n ± 0%  +0.92% (p=0.000 n=10)
SmallWriteStreamCompression/random-10      66.58n ± 0%   66.96n ± 0%  +0.57% (p=0.000 n=10)
StreamCompression-10                       73.04m ± 0%   73.65m ± 0%  +0.84% (p=0.000 n=10)
StreamDecompression-10                     10.10m ± 1%   10.01m ± 0%  -0.95% (p=0.002 n=10)
Compression-10                             66.05m ± 0%   67.22m ± 0%  +1.77% (p=0.000 n=10)
Decompression-10                           8.557m ± 0%   8.581m ± 0%  +0.28% (p=0.000 n=10)
geomean                                    108.6µ        109.2µ       +0.57%

                                         │     old      │                 new                 │
                                         │     B/s      │     B/s       vs base               │
BulkCompress-10                            98.44Mi ± 4%   97.58Mi ± 2%       ~ (p=0.481 n=10)
BulkDecompress-10                          130.9Mi ± 0%   132.2Mi ± 0%  +1.04% (p=0.000 n=10)
CtxCompression-10                          144.2Mi ± 0%   141.7Mi ± 0%  -1.76% (p=0.000 n=10)
CtxDecompression-10                        1.086Gi ± 0%   1.082Gi ± 0%  -0.38% (p=0.000 n=10)
SmallWriteStreamCompression/all-zeros-10   14.56Mi ± 0%   14.44Mi ± 0%  -0.85% (p=0.000 n=10)
SmallWriteStreamCompression/count-10       14.56Mi ± 0%   14.43Mi ± 0%  -0.88% (p=0.000 n=10)
SmallWriteStreamCompression/random-10      14.32Mi ± 0%   14.24Mi ± 0%  -0.57% (p=0.000 n=10)
StreamCompression-10                       130.2Mi ± 0%   129.1Mi ± 0%  -0.83% (p=0.000 n=10)
StreamDecompression-10                     941.2Mi ± 1%   950.2Mi ± 0%  +0.96% (p=0.002 n=10)
Compression-10                             144.0Mi ± 0%   141.5Mi ± 0%  -1.73% (p=0.000 n=10)
Decompression-10                           1.085Gi ± 0%   1.082Gi ± 0%  -0.28% (p=0.000 n=10)
geomean                                    125.6Mi        124.9Mi       -0.56%

                                         │         old         │                     new                      │
                                         │ compressed_bytes/op │ compressed_bytes/op  vs base                 │
SmallWriteStreamCompression/all-zeros-10           31.50µ ± 0%           31.50µ ± 0%       ~ (p=1.000 n=10)
SmallWriteStreamCompression/count-10               31.50µ ± 0%           31.50µ ± 0%       ~ (p=1.000 n=10)
SmallWriteStreamCompression/random-10               1.000 ± 0%            1.000 ± 0%       ~ (p=1.000 n=10) ¹
geomean                                            997.4µ                997.4µ       +0.00%
¹ all samples are equal
```